Rename to python-systemd-dbus to avoid conflicts
