+++ README
@@ -1,8 +1,8 @@
 ==============
-python-systemd
+python-systemd-dbus
 ==============
 
-python-systemd python wrapper for `systemd`_ system and session manager dbus
+python-systemd-dbus python wrapper for `systemd`_ system and session manager dbus
 interfaces.
 
 .. systemd: http://www.freedesktop.org/wiki/Software/systemd
@@ -10,7 +10,7 @@ interfaces.
 Basic usage
 ===========
 
->>> from systemd.manager import Manager
+>>> from systemd_dbus.manager import Manager
 >>> manager = Manager()
 
 # List all units
@@ -36,7 +36,7 @@ loaded active running
 
 # Let's stop crond
 >>> unit.stop('fail')
-<systemd.job.Job object at 0x7fa57ba03a90>
+<systemd_dbus.job.Job object at 0x7fa57ba03a90>
 
 # is crond running? why I stop it!!
 >>> print unit.properties.LoadState, unit.properties.ActiveState, unit.properties.SubState
@@ -54,7 +54,7 @@ loaded inactive dead
 
 # Let's start crond
 >>> unit.start('fail')
-<systemd.job.Job object at 0x7fa57ba03950>
+<systemd_dbus.job.Job object at 0x7fa57ba03950>
 
 # Remember we want o loop!
 >>> print unit.properties.LoadState, unit.properties.ActiveState, unit.properties.SubState
