+++ TODO
@@ -2,7 +2,7 @@
 TODO
 ====
 
-This is todo file for python-systemd.
+This is todo file for python-systemd-dbus.
 
 CODE
 ====
