+++ systemd_dbus/manager.py
@@ -1,14 +1,14 @@
 #
 # Copyright (c) 2010 Mandriva
 #
-# This file is part of python-systemd.
+# This file is part of python-systemd-dbus.
 #
-# python-systemd is free software; you can redistribute it and/or modify
+# python-systemd-dbus is free software; you can redistribute it and/or modify
 # it under the terms of the GNU Lesser General Public License as
 # published by the Free Software Foundation; either version 2.1 of
 # the License, or (at your option) any later version.
 #
-# python-systemd is distributed in the hope that it will be useful,
+# python-systemd-dbus is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU Lesser General Public License for more details.
@@ -21,10 +21,10 @@ import dbus
 import dbus.mainloop.glib
 dbus.mainloop.glib.DBusGMainLoop(set_as_default=True)
 
-from systemd.unit import Unit
-from systemd.job import Job
-from systemd.property import Property
-from systemd.exceptions import SystemdError
+from systemd_dbus.unit import Unit
+from systemd_dbus.job import Job
+from systemd_dbus.property import Property
+from systemd_dbus.exceptions import SystemdError
 
 class Manager(object):
     """Abstraction class to org.freedesktop.systemd1.Manager interface"""
@@ -93,7 +93,7 @@ class Manager(object):
         
         @raise SystemdError: Raised when no job is found with the given ID.
         
-        @rtype: systemd.job.Job
+        @rtype: systemd_dbus.job.Job
         """
         try:
             job_path = self.__interface.GetJob(ID)
@@ -109,7 +109,7 @@ class Manager(object):
         
         @raise SystemdError: Raised when no unit is found with the given name.
         
-        @rtype: systemd.unit.Unit
+        @rtype: systemd_dbus.unit.Unit
         """
         try:
             unit_path = self.__interface.GetUnit(name)
@@ -125,7 +125,7 @@ class Manager(object):
         
         @raise SystemdError: Raised when no unit with that PID is found.
         
-        @rtype: systemd.unit.Unit
+        @rtype: systemd_dbus.unit.Unit
         """
         try:
             unit_path = self.__interface.GetUnitByPID(pid)
@@ -157,7 +157,7 @@ class Manager(object):
         
         @raise SystemdError: Raised when no unit is found with the given name.
         
-        @rtype: L{systemd.job.Job}
+        @rtype: L{systemd_dbus.job.Job}
         """
         try:
             self.__interface.KillUnit(name, who, mode, signal)
@@ -171,7 +171,7 @@ class Manager(object):
         @raise SystemdError, IndexError: Raised when dbus error or index error
         is raised.
         
-        @rtype: A tuple of L{systemd.unit.Job}
+        @rtype: A tuple of L{systemd_dbus.unit.Job}
         """
         try:
             jobs = []
@@ -187,7 +187,7 @@ class Manager(object):
         @raise SystemdError: Raised when dbus error or index error
         is raised.
         
-        @rtype: A tuple of L{systemd.unit.Unit}
+        @rtype: A tuple of L{systemd_dbus.unit.Unit}
         """
         try:
             units = []
@@ -204,7 +204,7 @@ class Manager(object):
         
         @raise SystemdError: Raised when no unit is found with the given name.
         
-        @rtype: L{systemd.unit.Unit}
+        @rtype: L{systemd_dbus.unit.Unit}
         """
         try:
             unit_path = self.__interface.LoadUnit(name)
@@ -245,7 +245,7 @@ class Manager(object):
         
         @raise SystemdError: Raised when no unit is found with the given name.
         
-        @rtype: L{systemd.job.Job}
+        @rtype: L{systemd_dbus.job.Job}
         """
         try:
             job_path = self.__interface.ReloadOrRestartUnit(name, mode)
@@ -262,7 +262,7 @@ class Manager(object):
         
         @raise SystemdError: Raised when no unit is found with the given name.
         
-        @rtype: L{systemd.job.Job}
+        @rtype: L{systemd_dbus.job.Job}
         """
         try:
             job_path = self.__interface.ReloadOrTryRestartUnit(name, mode)
@@ -280,7 +280,7 @@ class Manager(object):
         @raise SystemdError: Raised when no unit is found with the given name or
         mode is not corret.
         
-        @rtype: L{systemd.job.Job}
+        @rtype: L{systemd_dbus.job.Job}
         """
         try:
             job_path = self.__interface.ReloadUnit(name, mode)
@@ -309,7 +309,7 @@ class Manager(object):
         
         @raise SystemdError: Raised when no unit is found with the given name.
         
-        @rtype: L{systemd.job.Job}
+        @rtype: L{systemd_dbus.job.Job}
         """
         try:
             job_path = self.__interface.RestartUnit(name, mode)
@@ -332,7 +332,7 @@ class Manager(object):
         
         @raise SystemdError: Raised when no unit is found with the given name.
         
-        @rtype: L{systemd.job.Job}
+        @rtype: L{systemd_dbus.job.Job}
         """
         try:
             job_path = self.__interface.StartUnit(name, mode)
@@ -350,7 +350,7 @@ class Manager(object):
         
         @raise SystemdError: Raised when no unit is found with the given name.
         
-        @rtype: L{systemd.job.Job}
+        @rtype: L{systemd_dbus.job.Job}
         """
         try:
             job_path = self.__interface.StartUnitReplace(old_unit, new_unit, mode)
@@ -367,7 +367,7 @@ class Manager(object):
         
         @raise SystemdError: Raised when no unit is found with the given name.
         
-        @rtype: L{systemd.job.Job}
+        @rtype: L{systemd_dbus.job.Job}
         """
         try:
             job_path = self.__interface.StopUnit(name, mode)
@@ -392,7 +392,7 @@ class Manager(object):
         @raise SystemdError: Raised when no unit is found with the given name or
         mode is invalid.
         
-        @rtype: L{systemd.job.Job}
+        @rtype: L{systemd_dbus.job.Job}
         """
         try:
             job_path = self.__interface.TryRestartUnit(name, mode)
