+++ systemd_dbus/snapshot.py
@@ -1,14 +1,14 @@
 #
 # Copyright (c) 2010 Mandriva
 #
-# This file is part of python-systemd.
+# This file is part of python-systemd-dbus.
 #
-# python-systemd is free software; you can redistribute it and/or modify
+# python-systemd-dbus is free software; you can redistribute it and/or modify
 # it under the terms of the GNU Lesser General Public License as
 # published by the Free Software Foundation; either version 2.1 of
 # the License, or (at your option) any later version.
 #
-# python-systemd is distributed in the hope that it will be useful,
+# python-systemd-dbus is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU Lesser General Public License for more details.
@@ -21,11 +21,11 @@ import dbus
 import dbus.mainloop.glib
 dbus.mainloop.glib.DBusGMainLoop(set_as_default=True)
 
-from systemd.property import Property
-from systemd.exceptions import SystemdError
+from systemd_dbus.property import Property
+from systemd_dbus.exceptions import SystemdError
 
-class Mount(object):
-    """Abstraction class to org.freedesktop.systemd1.Mount interface"""
+class Snapshot(object):
+    """Abstraction class to org.freedesktop.systemd1.Snapshot interface"""
     def __init__(self, unit_path):
         self.__bus = dbus.SystemBus()
 
@@ -35,7 +35,7 @@ class Mount(object):
 
         self.__interface = dbus.Interface(
             self.__proxy,
-            'org.freedesktop.systemd1.Mount',)
+            'org.freedesktop.systemd1.Snapshot',)
 
         self.__properties_interface = dbus.Interface(
             self.__proxy,
@@ -57,3 +57,9 @@ class Mount(object):
         for key, value in properties.items():
             setattr(attr_property, key, value)
         setattr(self, 'properties', attr_property)
+
+    def remove(self):
+        try:
+            self.__interface.Remove()
+        except dbus.exceptions.DBusException, error:
+            raise SystemdError(error)
