+++ systemd_dbus/timer.py
@@ -1,14 +1,14 @@
 #
 # Copyright (c) 2010 Mandriva
 #
-# This file is part of python-systemd.
+# This file is part of python-systemd-dbus.
 #
-# python-systemd is free software; you can redistribute it and/or modify
+# python-systemd-dbus is free software; you can redistribute it and/or modify
 # it under the terms of the GNU Lesser General Public License as
 # published by the Free Software Foundation; either version 2.1 of
 # the License, or (at your option) any later version.
 #
-# python-systemd is distributed in the hope that it will be useful,
+# python-systemd-dbus is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU Lesser General Public License for more details.
@@ -21,21 +21,21 @@ import dbus
 import dbus.mainloop.glib
 dbus.mainloop.glib.DBusGMainLoop(set_as_default=True)
 
-from systemd.property import Property
-from systemd.exceptions import SystemdError
+from systemd_dbus.property import Property
+from systemd_dbus.exceptions import SystemdError
 
-class Job(object):
-    """Abstraction class to org.freedesktop.systemd1.Job interface"""
-    def __init__(self, job_path):
+class Timer(object):
+    """Abstraction class to org.freedesktop.systemd1.Timer interface"""
+    def __init__(self, unit_path):
         self.__bus = dbus.SystemBus()
+
         self.__proxy = self.__bus.get_object(
             'org.freedesktop.systemd1',
-            job_path,
-        )
+            unit_path,)
+
         self.__interface = dbus.Interface(
             self.__proxy,
-            'org.freedesktop.systemd1.Job',
-        )
+            'org.freedesktop.systemd1.Timer',)
 
         self.__properties_interface = dbus.Interface(
             self.__proxy,
@@ -57,9 +57,3 @@ class Job(object):
         for key, value in properties.items():
             setattr(attr_property, key, value)
         setattr(self, 'properties', attr_property)
-
-    def cancel(self):
-        try:
-            self.__interface.Cancel()
-        except dbus.exceptions.DBusException, error:
-            raise SystemdError(error)
