+++ systemd_dbus/unit.py
@@ -1,14 +1,14 @@
 #
 # Copyright (c) 2010 Mandriva
 #
-# This file is part of python-systemd.
+# This file is part of python-systemd-dbus.
 #
-# python-systemd is free software; you can redistribute it and/or modify
+# python-systemd-dbus is free software; you can redistribute it and/or modify
 # it under the terms of the GNU Lesser General Public License as
 # published by the Free Software Foundation; either version 2.1 of
 # the License, or (at your option) any later version.
 #
-# python-systemd is distributed in the hope that it will be useful,
+# python-systemd-dbus is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU Lesser General Public License for more details.
@@ -21,9 +21,9 @@ import dbus
 import dbus.mainloop.glib
 dbus.mainloop.glib.DBusGMainLoop(set_as_default=True)
 
-from systemd.property import Property
-from systemd.exceptions import SystemdError
-from systemd.job import Job
+from systemd_dbus.property import Property
+from systemd_dbus.exceptions import SystemdError
+from systemd_dbus.job import Job
 
 class Unit(object):
     """Abstraction class to org.freedesktop.systemd1.Unit interface"""
@@ -69,7 +69,7 @@ class Unit(object):
         
         @raise SystemdError: Raised when who, mode or signal are invalid.
         
-        @rtype: systemd.job.Job
+        @rtype: systemd_dbus.job.Job
         """
         try:
             self.__interface.KillUnit(who, mode, signal)
@@ -84,7 +84,7 @@ class Unit(object):
         
         @raise SystemdError: Raised when mode is invalid.
         
-        @rtype: systemd.job.Job
+        @rtype: systemd_dbus.job.Job
         """
         try:
             job_path = self.__interface.Reload(mode)
@@ -102,7 +102,7 @@ class Unit(object):
         
         @raise SystemdError: Raised when mode is invalid.
         
-        @rtype: systemd.job.Job
+        @rtype: systemd_dbus.job.Job
         """
         try:
             job_path = self.__interface.ReloadOrRestart(mode)
@@ -118,7 +118,7 @@ class Unit(object):
         
         @raise SystemdError: Raised when mode is invalid.
         
-        @rtype: systemd.job.Job
+        @rtype: systemd_dbus.job.Job
         """
         try:
             job_path = self.__interface.ReloadOrTryRestart(mode)
@@ -140,7 +140,7 @@ class Unit(object):
         
         @raise SystemdError: Raised when mode is invalid.
         
-        @rtype: systemd.job.Job
+        @rtype: systemd_dbus.job.Job
         """
         try:
             job_path = self.__interface.Restart(mode)
@@ -156,7 +156,7 @@ class Unit(object):
         
         @raise SystemdError: Raised when mode is invalid.
         
-        @rtype: systemd.job.Job
+        @rtype: systemd_dbus.job.Job
         """
         try:
             job_path = self.__interface.Start(mode)
@@ -172,7 +172,7 @@ class Unit(object):
         
         @raise SystemdError: Raised when mode is invalid.
         
-        @rtype: systemd.job.Job
+        @rtype: systemd_dbus.job.Job
         """
         try:
             job_path = self.__interface.Stop(mode)
@@ -188,7 +188,7 @@ class Unit(object):
         
         @raise SystemdError: Raised when mode is invalid.
         
-        @rtype: L{systemd.job.Job}
+        @rtype: L{systemd_dbus.job.Job}
         """
         try:
             job_path = self.__interface.TryRestart(mode)
